+++ src/org/ggp/base/player/request/grammar/StopRequest.java
@@ -24,7 +24,7 @@ class StopRequest(Request):
     def getMatchId():  # String
         return matchId;
 
-    def String process(long receptionTime)
+    def String process(int receptionTime)
 	{
         // First, check to ensure that this stop request is for the match
         // we're currently playing. If we're not playing a match, or we're
